Nothing Changed
Files: Makefile, kernel.cpp, types.h

+++ types.h
@@ -22,8 +22,8 @@ typedef unsigned char u8; //8 bit unsigned integer (1 byte)
 typedef signed short s16; //16 bit integer (2 byte)
 typedef unsigned short u16; //16 bii unsigned integer (2 byte)
 
-typedef signed long s32; //32 bit integer (4bytes)
-typedef unsigned long u32; //32 bit unsigned integer (4bytes)
+typedef signed int s32; //32 bit integer (4bytes)
+typedef unsigned int u32; //32 bit unsigned integer (4bytes)
 
 typedef signed long long int s64; //64 bit integer (8bytes)
 typedef unsigned long long int u64; //64bit unsigned integer (8bytes)
